Move set_font_size()
