--- foo_ui_columns/font_notify.h
@@ -1,17 +1,0 @@
-#ifndef _COLUMNS_FONT_H_
-#define _COLUMNS_FONT_H_
-
-/*!
- * \file font_notify.h
- *
- * \author musicmusic
- * \date 1 March 2015
- *
- * Handles font changes for main UI components
- */
-
-#include "stdafx.h"
-
-void set_font_size(bool up);
-
-#endif
