+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -389,7 +389,6 @@
     <ClCompile Include=".\callbacks_config.cpp" />
     <ClCompile Include=".\commandline.cpp" />
     <ClCompile Include=".\common.cpp" />
-    <ClCompile Include=".\font_notify.cpp" />
     <ClCompile Include=".\menu_items.cpp" />
     <ClCompile Include=".\splitter_window.cpp" />
     <ClCompile Include=".\splitter_window_wndproc.cpp" />
@@ -491,7 +490,6 @@
     <ClInclude Include=".\rebar.h" />
     <ClInclude Include=".\common.h" />
     <ClInclude Include=".\extern.h" />
-    <ClInclude Include=".\font_notify.h" />
     <ClInclude Include=".\menu_items.h" />
     <ClInclude Include=".\status_bar.h" />
     <ClInclude Include=".\stdafx.h" />
