+++ foo_ui_columns/user_interface_impl.cpp
@@ -1,6 +1,5 @@
 #include "stdafx.h"
 #include "status_pane.h"
-#include "font_notify.h"
 #include "status_bar.h"
 #include "notification_area.h"
